Removed all V$ project and solution files from LeadwerksTemplate project.
Due to local references for the Leadwerks API I am removing the solution and project files from the repository. These files should be local to the user. This will allow users of the project to create a new Leadwerks project through the Project Manager and replace their projects source with the templates source files.