DbCommand no longer returns a record with numAffectedRows.
The Query method on a select runs until all record results are read into a result set however, the execution loop runs one more time before terminating, causing the numAffected record to be appended to the returned result set. The affectedRows record is now only added when the read result set is empty.